Conflicts:
	.vs/gameproject/v14/.suo
	Debug/gameproject.exe
	Debug/gameproject.ilk
	Debug/gameproject.pdb
	gameproject.sdf
	gameproject/Debug/Bomber.obj
	gameproject/Debug/Building.obj
	gameproject/Debug/Button.obj
	gameproject/Debug/CommandQueue.obj
	gameproject/Debug/Game.obj
	gameproject/Debug/GameObject.obj
	gameproject/Debug/InputHandler.obj
	gameproject/Debug/MenuButton.obj
	gameproject/Debug/Music.obj
	gameproject/Debug/Player.obj
	gameproject/Debug/PlayerButton.obj
	gameproject/Debug/PlayerObject.obj
	gameproject/Debug/ReadInput.obj
	gameproject/Debug/Resource.obj
	gameproject/Debug/Soldier.obj
	gameproject/Debug/Sound.obj
	gameproject/Debug/Terrain.obj
	gameproject/Debug/Unit.obj
	gameproject/Debug/UnitButton.obj
	gameproject/Debug/V2Functions.obj
	gameproject/Debug/gameproject.log
	gameproject/Debug/gameproject.tlog/CL.read.1.tlog
	gameproject/Debug/gameproject.tlog/CL.write.1.tlog
	gameproject/Debug/gameproject.tlog/link.read.1.tlog
	gameproject/Debug/main.obj
	gameproject/Debug/vc140.idb
	gameproject/Debug/vc140.pdb
	gameproject/Enums.h
	gameproject/level.txt

+++ gameproject/Bomber.cpp
@@ -3,7 +3,7 @@
 #include <SFML/Graphics.hpp>
 #include "Bomber.h"
 
-Bomber::Bomber(textureID id, const textureHolder& textures, sf::Vector2f pos, sf::Color color, int scale) :
+Bomber::Bomber(textureID id, const textureHolder& textures, sf::Vector2f pos, sf::Color color, float scale) :
 	Unit{ id, textures, pos, color, scale }
 {
 	hitpoints = 150;
