+++ gameproject/Bomber.h
@@ -8,7 +8,7 @@
 
 class Bomber : public Unit {
 public:
-	Bomber(textureID id, const textureHolder& textures, sf::Vector2f pos, sf::Color color, int scale);
+	Bomber(textureID id, const textureHolder& textures, sf::Vector2f pos, sf::Color color, float scale);
 };
 
 #endif
